+++ src/operations/get_category_colors_operation.js
@@ -1,7 +1,7 @@
 // src/operations/get_category_colors_operation.js
 
 import { BaseOperation } from 'base_classes/base_operation';
-import { Categories } from 'sheet_data/categories'; // 新しくimport
+import { Categories } from 'sheet_data/bound_sheet/categories';
 
 /**
  * 'categories'シートからカテゴリと色の対応を取得するオペレーション。
