+++ src/operations/get_config_operation.js
@@ -1,5 +1,5 @@
 import { BaseOperation } from "base_classes/base_operation";
-import { Config } from "sheet_data/config";
+import { Config } from "sheet_data/bound_sheet/config";
 
 export class GetConfigOperation extends BaseOperation {
   _operation() {
